requirements.txt
requests==2.28.2
google-crc32c==1.5.0
google-auth==2.16.2
google-auth-oauthlib==0.4.6
oauth2client==4.1.3
oauthlib==3.2.2
requests-oauthlib==1.3.1
google-cloud-bigquery==3.6.0
google-cloud-secret-manager==2.10.0
google-api-python-client==2.80.0
google-cloud-storage==2.7.0
